[wip] Cleanup geotrace branch on github
This updates the branch on github to reflect the version of
CLUBB used currently in iCAM6. There's a version mismatch here
as this has been partially rebased to a newer version of CLUBB,
therefore, many of the new modules have been commented out as
they do not fully work with the water tracer/isotope code yet.